+++ src/model/comments.php
@@ -1,49 +1,73 @@
 
 <?php 
-require ("model/crudmodel.php");
+require_once("config/bootstrap.php");
+require_once(WORKING_PATH."src/model/crudmode.php");
+
 class Comment extends Crudmodel{
     // define the class properties
 
 public $id = null;
+public $email = "";
 public $dynamicpages_id = null;
 public $comments= "";
 
 public function __construct($data=array()){
 if(isset($data['id']))$this->id=int($data['id']);
+if(isset($data['email']))$this->email=filter_var($data['email'],FILTER_VALIDATE_EMAIL);
 if(isset($data['dynamicpages_id']))$this->dynamicpagea_id=int($data['dynamicpages_id']);
 if(isset($data['comments']))$this->content=trim(stripslashes(htmlspecialchars(['comments'])));
 
 // define the class properties
-public function createComment(){
-    $this->insert("INSERT INTO comment(id,dynamicpages_id,comments)
-    VALUES(:id,:name,:dynamicpages_id , :comments)",["id"=>$this->id,"dynamicpages_id"=>$this->dynamicpages_id,
-    "comments"=>$this->comments,"created"=>$this->created]);
-    
+public function createComment($data=[]){
+    if(isset($data['email']) && isset($data['comments'])){
+    $dynamicpages = new Dynamicpages($_POST);
+    $url = $_GET['url'];
+    $result1 = $dynamicpages->displayAdynamicPage($url);
+    $this->dynamicpages_id = $result1['id'];
+    $result = $this->insert("INSERT INTO comments(id,dynamicpages_id,email,comments)
+    VALUES(:id,:name,:dynamicpages_id ,:email, :comments)",["id"=>$this->id,"dynamicpages_id"=>$this->dynamicpages_id,
+    "email"=>$this->email, "comments"=>$this->comments,"created"=>$this->created]);
+    }
+    if($result) return $result;
+    else{return false;}
         }
    
 
-public function readComment(){
-$this->select("SELECT * FROM comment");
+public function readComments(){
+$result = $this->select("SELECT * FROM comment");
+if($result)return $result;
+else(return false;)
 }
 
 public function readAComment($id){
-$this->select("SELECT* FROM comment WHERE id = :id", ["id"=>$this->id]);
-
+$thids->id = $id
+if(isset($id)){
+$result = $this->select("SELECT* FROM comment WHERE id = :id", ["id"=>$this->id]);
+}
+if($result)return $result;
+else{return false;}
 }
 
 
 
-public function updateComment($id,%data){
- $this->update("UPDATE cooment SET dynamicpages_id =:dynamicpages_id, comments = :comments WHERE id=:id",
- ["dynamicpages_id"=>$this->dynamicpages_id,"comments"=>%this->comments]);
+public function updateComment($id,$data =[]){
+    $this->id = $id;
+    if(isset($data['comments'])){
+$result = $this->update("UPDATE comment SET dynamicpages_id =:dynamicpages_id, email:email, comments = :comments WHERE id=:id",
+ ["dynamicpages_id"=>$this->dynamicpages_id,"email"=>$this->email, "comments"=>$this->comments,]);
 
+    }
+    if($result)return $result;
+    else(return false;)
 }
 
 public function deleteComment($id){
-        $this->delete("DELETE* FROM comment WHERE id=:id",["id"=>$this->id]);
+    $this->id = $id;
+    if(isset($id)){
+     $this->delete("DELETE* FROM comment WHERE id=:id",["id"=>$this->id]);
+    }
+      if($result)return $result:
+        else(return false;) 
+       }
        
-       
-           }
-       
-       
-}   
+    }
